Add support for serializing i128/u128 values
The crate currently fails serialization of 128 bit wide integers. That
is very unfortunate, because `SystemTime` values converted to
milliseconds, for example, are natively expressed as objects of such a
type.
`itoa` already supports such wide integer types, guarded by the i128
feature. With this change we introduce a similar feature and add support
for such a conversion.